Crate parquet_concat

Binary that concatenates the column data of one or more parquet files

Install

parquet-concat can be installed using cargo:

cargo install parquet --features=cli

After this parquet-concat should be available:

parquet-concat out.parquet a.parquet b.parquet

The binary can also be built from the source code and run as follows:

cargo run --features=cli --bin parquet-concat out.parquet a.parquet b.parquet

Note: this does not currently support preserving the page index or bloom filters

Structs

Args 🔒

Concatenates one or more parquet files

Functions

main 🔒